se crean tablas de bodegas se crea y elimina bodega

+++ database/migrations/2021_08_02_153028_create_product_warehouses_table.php
@@ -17,7 +17,7 @@ class CreateProductWarehousesTable extends Migration
             $table->increments('id');
             $table->integer('quantity')->nullable();
 
-            $table->integer('product_id')->unsigned()->nullable();
+            $table->bigInteger('product_id')->unsigned()->nullable();
             $table->foreign('product_id')->references('id')->on('products');
 
             $table->integer('warehouse_id')->unsigned()->nullable();
